+++ api/helpers/recipients/add.js
@@ -59,8 +59,12 @@ module.exports = {
       }
 
       sails.log.silly(`sails.models.status: ${status}`)
+      var answerCalls = false
+      var find = inputs.host
 
       // If this is a computers recipient, see if it's in the Hosts table. If so, use that as the label instead of the provided label.
+      // Also, use generic `computer-uniqueString` as host instead of actual host for security purposes.
+      // Also, determine if this recipient is allowed to answer calls.
       if (inputs.group === 'computers') {
         var host = await sails.models.hosts.find({ host: inputs.host }).limit(1)
           .tolerate((err) => {
@@ -68,16 +72,18 @@ module.exports = {
           })
         if (host && typeof host[0] !== 'undefined') {
           inputs.label = host[0].friendlyname
+          answerCalls = host[0].answerCalls && host[0].authorized
+          find = `computer-${sh.unique(inputs.host + sails.config.custom.hostSecret)}`
         }
       }
 
       // Get or create the recipient entry
-      var recipient = await sails.models.recipients.findOrCreate({ host: inputs.host }, { host: inputs.host, device: inputs.device, group: inputs.group, label: inputs.label, status: status, time: moment().toISOString(true) })
+      var recipient = await sails.models.recipients.findOrCreate({ host: find }, { host: find, device: inputs.device, group: inputs.group, label: inputs.label, status: status, answerCalls: answerCalls, time: moment().toISOString(true) })
 
       sails.log.silly(`sails.models.recipients record: ${recipient}`)
 
       // Search to see if any changes are made to the recipient; we only want to update if there is a change.
-      var criteria = { host: inputs.host, group: inputs.group, status: status, device: inputs.device }
+      var criteria = { host: find, group: inputs.group, status: status, device: inputs.device, answerCalls: answerCalls }
       var updateIt = false
       for (var key in criteria) {
         if (Object.prototype.hasOwnProperty.call(criteria, key)) {
@@ -92,7 +98,7 @@ module.exports = {
       }
       if (updateIt) {
         sails.log.verbose(`Updating recipient as it has changed.`)
-        await sails.models.recipients.update({ host: inputs.host }, { host: inputs.host, group: inputs.group, device: inputs.device, status: status, time: moment().toISOString(true) }).fetch()
+        await sails.models.recipients.update({ host: find }, { host: find, group: inputs.group, device: inputs.device, status: status, answerCalls: answerCalls, time: moment().toISOString(true) }).fetch()
       }
 
       // Put the socket ID in memory